+++ app/lib/logger.js
@@ -1,16 +1,1 @@
-const bunyan = require('bunyan');
-
-const log = bunyan.createLogger({ name: 'pharmacy-data-etl' });
-
-function info(message) {
-  log.info(message);
-}
-
-function error(message, err) {
-  log.error(message, err);
-}
-
-module.exports = {
-  info,
-  error,
-};
+module.exports = require('nhsuk-bunyan-logger')('pharmacy-data-etl');